Added API to run matilda in native process using executable
