--- matilda/exceptions/matilda_java_process_not_started_exception.py
@@ -1,3 +1,0 @@
-class MatildaJavaProcessNotStartedException(Exception):
-    def __init__(self, exit_code: int, stderr: bytes):
-        super().__init__(f"Matilda java process could not be started. Exit code: {exit_code} ({stderr})")
