+++ matilda/exceptions/matilda_process_not_started_exception.py
@@ -1,0 +1,3 @@
+class MatildaProcessNotStartedException(Exception):
+    def __init__(self, exit_code: int, stderr: bytes):
+        super().__init__(f"Matilda process could not be started. Exit code: {exit_code} ({stderr})")
